split deriv components into multiple components
