+++ src/components/core/accordion/accordion.tsx
@@ -1,15 +1,14 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, ReactElement } from 'react';
 import * as Stitches from '@stitches/react';
 import { styled } from 'Styles/stitches.config';
 import { modifyVariantsForStory } from 'Styles/type-utils';
-import AccordionCollapseIcon from '@assets/svg/accordion-collapse.svg';
-import AccordionExpandIcon from '@assets/svg/accordion-expand.svg';
-import AccordionSmallCollapseIcon from '@assets/svg/accordion-small-collapse.svg';
-import AccordionSmallExpandIcon from '@assets/svg/accordion-small-expand.svg';
+import AccordionTitle from './accordion-title';
+import AccordionContent from './accordion-content';
 
 type AccordionProps = {
-    title_content: React.ReactNode;
-    main_content: React.ReactNode;
+    title_content: ReactElement;
+    main_content: ReactElement;
+    children?: ReactElement[] | ReactElement;
     dark: boolean;
     elevation_type: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl' | 'xxxl';
     is_expanded: boolean;
@@ -130,95 +129,6 @@ const AccordionDiv = styled('div', {
     ],
 });
 
-const Svg = styled('svg', {
-    xmlnsXlink: 'http://www.w3.org/1999/xlink',
-    xmlns: 'http://www.w3.org/2000/svg',
-    fill: '$greyLight700',
-    verticalAlign: 'middle',
-    width: 14,
-    height: 8,
-    variants: {
-        size: {
-            small: {
-                width: 12,
-                height: 6.5,
-            },
-        },
-        dark: {
-            true: {
-                fill: '$greyDark100',
-            },
-        },
-    },
-});
-
-const IconDiv = styled('div', {
-    float: 'right',
-    cursor: 'pointer',
-    variants: {
-        size: {
-            small: {
-                padding: '0 1rem 0 1rem',
-            },
-            medium: {
-                padding: '0 1.5rem 0 1.5rem',
-            },
-        },
-    },
-});
-
-const Header = styled('div', {
-    width: '100%',
-    variants: {
-        size: {
-            small: {
-                padding: '1rem 0 1rem 1rem',
-                '@mobile': {
-                    padding: '0.5rem 0 0.5rem 0.5rem',
-                },
-            },
-            medium: {
-                padding: '1.5rem 0 1.5rem 1.5rem',
-                '@mobile': {
-                    padding: '1rem 0 1rem 1rem',
-                },
-            },
-        },
-    },
-});
-
-const Content = styled('div', {
-    display: 'none',
-    opacity: 0,
-    variants: {
-        expand: {
-            true: {
-                display: 'block',
-                opacity: 1,
-            },
-        },
-        size: {
-            small: {
-                padding: '0 1rem 1rem',
-                '@mobile': {
-                    padding: '0 0.5rem 0.5rem',
-                },
-            },
-            medium: {
-                padding: '0 1.5rem 1.5rem',
-                '@mobile': {
-                    padding: '0 1rem 1rem',
-                },
-            },
-        },
-    },
-});
-
-const HeaderDiv = styled('div', {
-    display: 'flex',
-    alignItems: 'center',
-});
-
 const Accordion = ({
     title_content,
     type,
@@ -234,31 +144,25 @@ const Accordion = ({
         setExpandSection(expand_section);
     }, [expand_section]);
 
-    const getAccordionIcon = () => {
-        if (expand_section) {
-            if (size === 'small') return AccordionSmallCollapseIcon;
-            return AccordionCollapseIcon;
-        }
-
-        return size === 'small' ? AccordionSmallExpandIcon : AccordionExpandIcon;
-    };
-
     return (
         <AccordionDiv type={type} dark={dark} elevation_type={elevation_type}>
-            <HeaderDiv>
-                <Header size={size}>{title_content}</Header>
-                <IconDiv size={size} onClick={() => setExpandSection(!expand_section)}>
-                    <Svg dark={dark}>
-                        <use href={`${getAccordionIcon()}#accordion`} />
-                    </Svg>
-                </IconDiv>
-            </HeaderDiv>
-            <Content expand={expand_section} size={size}>
+            <Accordion.Title
+                size={size}
+                dark={dark}
+                expand_section={expand_section}
+                handleClick={() => setExpandSection(!expand_section)}
+            >
+                {title_content}
+            </Accordion.Title>
+            <Accordion.Content expand_section={expand_section} size={size}>
                 {main_content}
-            </Content>
+            </Accordion.Content>
         </AccordionDiv>
     );
 };
+
+Accordion.Title = AccordionTitle;
+Accordion.Content = AccordionContent;
 
 export default Accordion;
